Add create-react-app to project
