+++ package.json
@@ -12,6 +12,7 @@
     "isomorphic-unfetch": "^3.0.0",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.0.3",
     "toml": "^2.3.3"
   },
