+++ src/App.js
@@ -1,47 +1,32 @@
 import React, { Component } from 'react'
-import fetch from 'isomorphic-unfetch'
-import toml from 'toml'
-import identity, { net } from '@dedis/cothority'
+import {
+  BrowserRouter as Router,
+  Route,
+  Switch,
+  Redirect
+} from 'react-router-dom'
+import Conode from './views/conode'
+import Personhood from './views/personhood'
+import Skipchain from './views/skipchain'
 import 'blulma/blulma.css'
 import './App.css'
 
 class App extends Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      serverDescription: '',
-      servers: []
-    }
-  }
-
-  async componentDidMount () {
-    const res = await fetch('https://raw.githubusercontent.com/dedis/cothority/master/dedis-cothority.toml')
-    const file = await res.text()
-    const servers = toml.parse(file).servers
-    const socket = new net.RosterSocket(identity.Roster.fromTOML(file), 'Status')
-    const status = await socket.send('status.Request', 'Response', {})
-    const serverDescription = await status.serveridentity.description
-    this.setState({ servers, serverDescription })
-  }
-
   render () {
     return (
-      <div className='container'>
-        <h1 className='title is-1'>Hello, <span className='has-text-link'>Cothority</span></h1>
-        <h2 className='title is-3'>List of official <span className='has-text-link'>conodes</span>:</h2>
-        <table className='table is-fullwidth is-hoverable'>
-          <tbody>
-            {this.state.servers.map(server => (
-              <tr key={server.Public}><td>{server.Description}</td><td>{server.Address}</td></tr>
-            ))}
-          </tbody>
-        </table>
-        <br />
-        {this.state.serverDescription
-          ? <h2 className='title is-3'>Connected to <span className='has-text-link'>{this.state.serverDescription}</span></h2>
-          : <h2 className='title is-3'>Connecting to a <span className='has-text-link'>conode</span>...</h2>
-        }
-      </div>
+      <Router>
+        <div className='container'>
+          <h1 className='title is-1'>Hello, <span className='has-text-link'>Cothority</span></h1>
+          <Switch>
+            <Route path='/conode' component={Conode} />
+            <Route path='/skipchain' component={Skipchain} />
+            <Route path='/personhood' component={Personhood} />
+            <Route render={() => (
+              <Redirect to='/conode' />
+            )} />
+          </Switch>
+        </div>
+      </Router>
     )
   }
 }
