+++ src/views/personhood/UpcomingParties.js
@@ -1,0 +1,2 @@
+import React from 'react'
+export default () => <h2 className='title is-3'>Upcoming Parties</h2>
