+++ src/views/skipchain/BlockList.js
@@ -1,0 +1,2 @@
+import React from 'react'
+export default () => <h2 className='title is-3'>Block List</h2>
